Modifying the test to delete and verify whether bucket is deleted. Also added conftest
